Remove spellbook from search view, create own Link
