+++ src/App/App.js
@@ -3,6 +3,7 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import Homepage from '../Homepage/Homepage';
 import SpellSearch from '../SpellSearch/SpellSearch';
+import SpellBook from '../SpellBook/SpellBook';
 
 const App = () => {
   return (
@@ -10,6 +11,7 @@ const App = () => {
       <Switch>
         <Route exact path="/" component={Homepage} />
         <Route exact path="/spells/:pcClass" component={SpellSearch} />
+        <Route exact path="/spellbook" component={SpellBook} />
       </Switch>
     </div>
   );
