+++ src/SpellSearch/SpellSearch.js
@@ -5,45 +5,48 @@ import { useEffect, useState } from 'react';
 import * as API from '../APIcalls';
 import SpellCard from '../SpellCard/SpellCard';
 import SpellDetails from '../SpellDetails/SpellDetails';
-import SpellBook from '../SpellBook/SpellBook';
+// import SpellBook from '../SpellBook/SpellBook';
 
 const SpellSearch = () => {
   const [spells, setSpells] = useState([]);
   const [spellDetails, setSpellDetails] = useState([]);
   const [displayedSpell, setDisplayedSpell] = useState(null);
-  const [spellBook, setSpellBook] = useState([]);
+  // const [spellBook, setSpellBook] = useState([]);
   const searchCriteria = useParams().pcClass;
 
-  const fetchAllClassSpells = () => {
+  // const fetchAllClassSpells = () => {
+  //   API.fetchSpells(searchCriteria).then((data) => {
+  //     setSpells(data.results);
+  //   });
+  // };
+
+  // const sortSpells = (spells) => {
+  //   return spells.sort(
+  //     (orderedSpell, spell) => orderedSpell.level - spell.level
+  //   );
+  // };
+
+  // const loadSpellBookSpells = () => {
+  //   let spellBookSpells = [];
+  //   Object.values(localStorage).forEach((spell) =>
+  //     API.fetchSpellDetails(spell).then((data) => {
+  //       spellBookSpells.push(data);
+  //       if (localStorage.length === spellBookSpells.length) {
+  //         setSpellBook(sortSpells(spellBookSpells));
+  //       }
+  //     })
+  //   );
+  // };
+
+  useEffect(() => {
     API.fetchSpells(searchCriteria).then((data) => {
       setSpells(data.results);
     });
-  };
-
-  const sortSpells = (spells) => {
-    return spells.sort(
-      (orderedSpell, spell) => orderedSpell.level - spell.level
-    );
-  };
-
-  const loadSpellBookSpells = () => {
-    let spellBookSpells = [];
-    Object.values(localStorage).forEach((spell) =>
-      API.fetchSpellDetails(spell).then((data) => {
-        spellBookSpells.push(data);
-        if (localStorage.length === spellBookSpells.length) {
-          setSpellBook(sortSpells(spellBookSpells));
-        }
-      })
-    );
-  };
-
-  useEffect(() => {
-    if (searchCriteria === 'spellbook') {
-      loadSpellBookSpells();
-    } else {
-      fetchAllClassSpells();
-    }
+    // if (searchCriteria === 'spellbook') {
+    //   loadSpellBookSpells();
+    // } else {
+    // fetchAllClassSpells();
+    // }
   }, []);
 
   useEffect(() => {
@@ -72,35 +75,34 @@ const SpellSearch = () => {
     );
   };
 
-  if (searchCriteria === 'spellbook') {
-    return (
-      <section>
-        <SpellBook spellBook={spellBook} />
-      </section>
-    );
-  } else {
-    return (
-      <section className="search-wrapper">
-        <div className="all-spell-wrapper">
-          {spellDetails && (
-            <div className="spell-cards-wrapper">
-              <h1 className="search-criteria-title">
-                Spell Scrolls for{' '}
-                {searchCriteria.charAt(0).toUpperCase() +
-                  searchCriteria.slice(1)}
-              </h1>
-              {spellDetails.map((spell) => createSpellCard(spell))}{' '}
-            </div>
-          )}
-          <div className="spell-details-wrapper">
-            {displayedSpell && (
-              <SpellDetails spell={displayedSpell} view="scroll" />
-            )}
+  // if (searchCriteria === 'spellbook') {
+  //   return (
+  //     <section>
+  //       <SpellBook spellBook={spellBook} />
+  //     </section>
+  //   );
+  // } else {
+  return (
+    <section className="search-wrapper">
+      <div className="all-spell-wrapper">
+        {spellDetails && (
+          <div className="spell-cards-wrapper">
+            <h1 className="search-criteria-title">
+              Spell Scrolls for{' '}
+              {searchCriteria.charAt(0).toUpperCase() + searchCriteria.slice(1)}
+            </h1>
+            {spellDetails.map((spell) => createSpellCard(spell))}{' '}
           </div>
+        )}
+        <div className="spell-details-wrapper">
+          {displayedSpell && (
+            <SpellDetails spell={displayedSpell} view="scroll" />
+          )}
         </div>
-      </section>
-    );
-  }
+      </div>
+    </section>
+  );
 };
+// };
 
 export default SpellSearch;
